chore(restructure): split thumbnail -> thumbnails/{titles,images,descriptions}.py
Phase 1 step 6: separate the thumbnail workflow into three modules.
Each concern is < 100 lines and independently testable. Shim at the old
path preserves existing imports. Tests stay 62/62 green.
