+++ openshorts/thumbnails/descriptions.py
@@ -1,0 +1,65 @@
+"""YouTube description + chapter-marker generation from transcript segments."""
+
+from google import genai
+
+
+def generate_youtube_description(api_key, title, transcript_segments, language, video_duration):
+    """
+    Uses Gemini to generate a YouTube description with chapter markers from transcript segments.
+    Returns: { "description": "full description text with chapters" }
+    """
+    client = genai.Client(api_key=api_key)
+
+    # Format segments for the prompt
+    formatted_segments = []
+    for seg in transcript_segments:
+        start = seg.get("start", 0)
+        mins = int(start // 60)
+        secs = int(start % 60)
+        timestamp = f"{mins}:{secs:02d}"
+        formatted_segments.append(f"[{timestamp}] {seg.get('text', '').strip()}")
+
+    segments_text = "\n".join(formatted_segments)
+
+    # Format total duration
+    dur_mins = int(video_duration // 60)
+    dur_secs = int(video_duration % 60)
+    duration_str = f"{dur_mins}:{dur_secs:02d}"
+
+    prompt = f"""You are a YouTube SEO expert. Generate a complete YouTube video description for the following video.
+
+VIDEO TITLE: "{title}"
+VIDEO LANGUAGE: {language}
+VIDEO DURATION: {duration_str}
+
+TRANSCRIPT WITH TIMESTAMPS:
+{segments_text}
+
+REQUIREMENTS:
+1. Write the description in the SAME LANGUAGE as the video ({language})
+2. Start with a compelling 2-3 sentence summary/hook
+3. Add relevant CTAs (subscribe, like, comment)
+4. Generate YouTube CHAPTERS based on the transcript timestamps:
+   - First chapter MUST start at 0:00
+   - Minimum 3 chapters, each at least 10 seconds apart
+   - Chapter titles should be concise and descriptive
+   - Format: 0:00 Chapter Title
+   - Place chapters in their own section with a blank line before and after
+5. Add 5-10 relevant hashtags at the end
+6. Keep the total description under 5000 characters
+
+OUTPUT: Return ONLY the description text (no JSON wrapper, no markdown code blocks). The description should be ready to paste directly into YouTube."""
+
+    print("🤖 [Thumbnail] Generating YouTube description with chapters...")
+    response = client.models.generate_content(
+        model="gemini-2.5-flash",
+        contents=[prompt],
+    )
+
+    description = response.text.strip()
+    # Clean up any accidental markdown wrappers
+    if description.startswith("```"):
+        lines = description.split("\n")
+        description = "\n".join(lines[1:-1] if lines[-1].strip() == "```" else lines[1:])
+
+    return {"description": description}
