+++ openshorts/thumbnails/titles.py
@@ -1,0 +1,170 @@
+"""Gemini-driven viral title generation and conversational refinement."""
+
+import json
+import time
+from google import genai
+from google.genai import types
+
+
+def analyze_video_for_titles(api_key, video_path, transcript=None):
+    """
+    Transcribes a video and uses Gemini to suggest viral YouTube titles.
+    If transcript is provided, skips Whisper transcription.
+    Returns: { "titles": [...], "transcript_summary": "...", "language": "...", "segments": [...], "video_duration": ... }
+    """
+    if transcript is None:
+        from main import transcribe_video
+        print("🎬 [Thumbnail] Transcribing video...")
+        transcript = transcribe_video(video_path)
+    else:
+        print("🎬 [Thumbnail] Using pre-computed transcript (Whisper already done)...")
+
+    print("📤 [Thumbnail] Uploading video to Gemini...")
+    client = genai.Client(api_key=api_key)
+
+    file_upload = client.files.upload(file=video_path)
+    while True:
+        file_info = client.files.get(name=file_upload.name)
+        if file_info.state == "ACTIVE":
+            break
+        elif file_info.state == "FAILED":
+            raise Exception("Video processing failed by Gemini.")
+        time.sleep(2)
+
+    prompt = f"""You are a YouTube title expert who creates viral, click-worthy titles.
+
+Analyze this video and its transcript, then suggest 10 YouTube titles that would maximize CTR (click-through rate).
+
+TRANSCRIPT:
+{transcript['text']}
+
+RULES:
+- Titles must be under 70 characters
+- Use power words, curiosity gaps, and emotional triggers
+- Mix styles: how-to, listicle, story-driven, controversial, question-based
+- Make them specific to the actual content, not generic
+- Include numbers where appropriate
+- Consider the language of the video (detected: {transcript['language']})
+- Titles should be in the SAME LANGUAGE as the video transcript
+
+Also provide a brief summary of the video content (2-3 sentences).
+
+After generating all 10 titles, pick the TOP 2 you most recommend and explain concisely WHY (CTR potential, emotional hook, uniqueness, etc.). Reference them by their 0-based index in the titles array.
+
+OUTPUT JSON:
+{{
+    "titles": ["title1", "title2", ...],
+    "transcript_summary": "Brief summary of the video content...",
+    "language": "{transcript['language']}",
+    "recommended": [
+        {{"index": 0, "reason": "Why this title is best..."}},
+        {{"index": 3, "reason": "Why this title is second best..."}}
+    ]
+}}"""
+
+    print("🤖 [Thumbnail] Asking Gemini for title suggestions...")
+    response = client.models.generate_content(
+        model="gemini-2.5-flash",
+        contents=[file_upload, prompt],
+        config=types.GenerateContentConfig(
+            response_mime_type="application/json"
+        )
+    )
+
+    # Extract segments and duration from transcript for later use
+    segments = transcript.get("segments", [])
+    video_duration = segments[-1]["end"] if segments else 0
+
+    try:
+        text = response.text.strip()
+        if text.startswith("```json"):
+            text = text[7:]
+        if text.startswith("```"):
+            text = text[3:]
+        if text.endswith("```"):
+            text = text[:-3]
+        text = text.strip()
+
+        start_idx = text.find('{')
+        end_idx = text.rfind('}')
+        if start_idx != -1 and end_idx != -1:
+            text = text[start_idx:end_idx + 1]
+
+        result = json.loads(text)
+        result["transcript_summary"] = result.get("transcript_summary", "")
+        result["language"] = result.get("language", transcript["language"])
+        result["segments"] = segments
+        result["video_duration"] = video_duration
+        return result
+    except json.JSONDecodeError:
+        print(f"❌ [Thumbnail] Failed to parse titles JSON: {response.text}")
+        return {
+            "titles": ["Could not generate titles - please try again"],
+            "transcript_summary": transcript["text"][:500],
+            "language": transcript["language"],
+            "segments": segments,
+            "video_duration": video_duration
+        }
+
+
+def refine_titles(api_key, context, user_message, conversation_history=None):
+    """
+    Takes video context + user feedback and returns refined title suggestions.
+    """
+    client = genai.Client(api_key=api_key)
+
+    history_text = ""
+    if conversation_history:
+        for msg in conversation_history:
+            role = msg.get("role", "user")
+            history_text += f"\n{role.upper()}: {msg['content']}"
+
+    prompt = f"""You are a YouTube title expert. Based on the video context and the user's feedback, suggest 8 new refined YouTube titles.
+
+VIDEO CONTEXT:
+{context}
+
+CONVERSATION HISTORY:{history_text}
+
+USER'S NEW REQUEST:
+{user_message}
+
+RULES:
+- Titles must be under 70 characters
+- Incorporate the user's feedback/direction
+- Keep titles viral and click-worthy
+- If the user asks for a specific style, follow it
+- Titles should be in the same language as the original content
+
+OUTPUT JSON:
+{{
+    "titles": ["title1", "title2", ...]
+}}"""
+
+    response = client.models.generate_content(
+        model="gemini-2.5-flash",
+        contents=[prompt],
+        config=types.GenerateContentConfig(
+            response_mime_type="application/json"
+        )
+    )
+
+    try:
+        text = response.text.strip()
+        if text.startswith("```json"):
+            text = text[7:]
+        if text.startswith("```"):
+            text = text[3:]
+        if text.endswith("```"):
+            text = text[:-3]
+        text = text.strip()
+
+        start_idx = text.find('{')
+        end_idx = text.rfind('}')
+        if start_idx != -1 and end_idx != -1:
+            text = text[start_idx:end_idx + 1]
+
+        return json.loads(text)
+    except json.JSONDecodeError:
+        print(f"❌ [Thumbnail] Failed to parse refined titles: {response.text}")
+        return {"titles": ["Could not refine titles - please try again"]}
